requirements.txt
playwright==1.44.0
feedparser==6.0.11
jinja2==3.1.4
python-dateutil==2.9.0
